page

+++ app/controllers/experienceController.js
@@ -1,5 +1,7 @@
 const debug = require('debug')('app:api');
-const {Experience, validateExperience} = require('../models/experienceModel');
+const {Experience, validateExperience, experienceSchema} = require('../models/experienceModel');
+const mongoosePaginate = require('mongoose-paginate-v2');
+const {User} = require('../models/userModel');
 // TO-DO: pagination, null values
 
 
@@ -45,11 +47,17 @@ const like = async (req, res) => {
 // return all experiences 
 // TO-DO pagination-check from reem first
 const getExperiences = async (req, res) => {
+    experienceSchema.plugin(mongoosePaginate);
     
-    debug('get experiences');
+    try {
+        const result = await Experience.paginate({ page: req.query.pageNumber, limit: req.query.pageSize });
+        res.status(200).send(result);
+      } catch (error) {
+        res.status(404).render("err-response", { err: 404, msg: 'page not found :\( please check the URL and try again' });
+      }
+    };
 
-    res.send({experiences, end:false});
-};
+
 
 // get the user experiences
 const getExperience = async (req, res) => {
@@ -67,13 +75,26 @@ const getExperience = async (req, res) => {
 // only return the experience id
 const deleteExperience = async (req, res) => {
     const experienceID = req.params.id;
-    const result = await Experience.deleteOne(experienceID);
-    debug('delete an experience');
+    const experience = await Experience.findById(experienceID);
+    if(!experience)
+         res.status(404).render("err-response", { err: 404, msg: 'page not found :\( please check the URL and try again' });
+    
+    const userID = req.user._id;
+    const user = await User.findById(userID);
+    if(user.isAdmin == true || user.isAdmin == false && experience.user == userID){
+        const result = await Experience.deleteOne(experienceID);
+    }
+    else {
+        res.status(403).render("err-response", {err:403, msg: 'access denied... forbidden' });
+    }
+
     if (result.deletedCount === 1) {
     res.send({ id: experienceID });
     }
     else
     res.status(404).render("err-response", { err: 404, msg: 'Experience not deleted :\( please check the ID and try again' });
+
+    debug('delete an experience');
 };
 
 module.exports = {
